+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Additions
 - Sound when grabbing key.
 - Jump sound for moveable cubes.
-- Player trail.
+- Player emits a trail when moving.
+- Portal emits particles.
+- Player emits particles on death.
 
 ## Changes
 - Key hitbox changed, so no more wierd looking key grab.